requirements.txt
fpdf==1.7.2
Pillow==7.0.0
PyQt5==5.10
requests==2.23.0
selenium==3.141.0
xlrd==1.2.0
xlwt==1.3.0

